Related work items: #64, #69, #73, #76

+++ LegendaryDashboard.Application/Services/UserService/Implementations/UserService.cs
@@ -1,6 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Security.Cryptography;
+using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
@@ -8,8 +12,13 @@ using AutoMapper;
 using LegendaryDashboard.Application.Services.UserService.Interfaces;
 using LegendaryDashboard.Contracts.Contracts.User;
 using LegendaryDashboard.Contracts.Contracts.User.Requests;
+using LegendaryDashboard.Domain.Common;
 using LegendaryDashboard.Domain.Models;
 using LegendaryDashboard.Infrastructure.IRepositories;
+using LegendaryDashboard.Infrastructure.Options;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.Tokens;
 
 namespace LegendaryDashboard.Application.Services.UserService.Implementations
 {
@@ -17,14 +26,18 @@ namespace LegendaryDashboard.Application.Services.UserService.Implementations
     {
         private readonly IUserRepository _repository;
         private readonly IMapper _mapper;
+        private readonly IOptions<JwtOptions> _jwtOptions;
+        private readonly IHttpContextAccessor _accessor;
 
-        public UserService(IUserRepository repository, IMapper mapper)
+        public UserService(IUserRepository repository, IMapper mapper, IOptions<JwtOptions> jwtOptions, IHttpContextAccessor accessor)
         {
             _repository = repository;
             _mapper = mapper;
+            _jwtOptions = jwtOptions;
+            _accessor = accessor;
         }
         
-        public async Task Register(CreateUserRequest request, CancellationToken cancellationToken)
+        public async Task Register(RegisterUserRequest request, CancellationToken cancellationToken)
         {
             Match phoneValidation  = Regex.
                 Match(request.Phone,
@@ -45,13 +58,46 @@ namespace LegendaryDashboard.Application.Services.UserService.Implementations
             }
             
             var user = _mapper.Map<User>(request);
+            user.Role = RoleConstants.UserRole;
+            user.PasswordHash = Hashing.GetHash(request.Password);
             user.RegisterDate = DateTime.UtcNow;
             await _repository.Save(user, cancellationToken);
             
         }
+
+        public async Task<string> Login(LoginUserRequest request, CancellationToken cancellationToken)
+        {
+            var user = await _repository.GetByEmail(request.Email, cancellationToken);
+            if (!user.PasswordHash.Equals(Hashing.GetHash(request.Password)))
+            {
+                throw new Exception("Неверный email или пароль!");
+            }
+            var claims = new Claim[]
+            {
+                new Claim(ClaimTypes.NameIdentifier, user.Id.ToString()), 
+                new Claim(ClaimTypes.Role, user.Role)
+            };
+            var bytes = Encoding.ASCII.GetBytes(_jwtOptions.Value.Key);
+                var expires = DateTime.UtcNow.AddMinutes(90);
+                var securityKey = new SymmetricSecurityKey(bytes);
+                var credentials = new SigningCredentials(securityKey, SecurityAlgorithms.HmacSha256Signature);
+                var tokenHandler = new JwtSecurityTokenHandler();
+                var token = tokenHandler.CreateToken(new SecurityTokenDescriptor
+                {
+                    Expires = expires,
+                    SigningCredentials = credentials,
+                    Subject = new ClaimsIdentity(claims),
+                });
+                
+            return tokenHandler.WriteToken(token);
+        }
+
         public async Task Delete(int id, CancellationToken cancellationToken)
         {
-            await _repository.Delete(id, cancellationToken);
+            if (ClaimsPrincipalExtensions.IsAdminOrOwner(_accessor, id))
+            {
+                await _repository.Delete(id, cancellationToken);  
+            }
         }
         public async Task<IEnumerable<UserDto>> GetPaged(int offset, int limit, CancellationToken cancellationToken)
         {
@@ -77,6 +123,18 @@ namespace LegendaryDashboard.Application.Services.UserService.Implementations
         {
             var user = await _repository.GetByPhone(phone, cancellationToken);
             return _mapper.Map<UserDto>(user);
+        }
+    }
+
+    public static class Hashing
+    {
+        public static string GetHash(string str)
+        {
+            var strBytes = Encoding.ASCII.GetBytes(str);
+            var sha = new SHA256Managed();
+            var hash = sha.ComputeHash(strBytes);
+            var hashedStr = Encoding.Default.GetString(hash);
+            return hashedStr;
         }
     }
 }
